refactor: 지하철역 인수테스트
