+++ README.md
@@ -14,5 +14,6 @@
 - [X] 노선 이름 중복 금지 기능
   - [X] 인수 테스트
   - [X] 기능 구현 
-- [ ] 인수 테스트 리팩터링
+- [X] 인수 테스트 리팩터링
   - [X] 노선 인수 테스트 
+  - [X] 지하철역 인수 테스트
