+++ src/test/java/nextstep/subway/fixture/CommonFixture.java
@@ -1,0 +1,10 @@
+package nextstep.subway.fixture;
+
+import io.restassured.response.ExtractableResponse;
+import io.restassured.response.Response;
+
+public class CommonFixture {
+    public static String uri(ExtractableResponse<Response> response) {
+        return response.header("Location");
+    }
+}
